+++ src/view/main/movie.ts
@@ -4,7 +4,7 @@ import fetchMovies from '../../api/fetchMovies';
 import starFilledImage from '../../assets/images/star_filled.png';
 import movieStateMethod from '../../store/movieStore';
 import { renderMovieDetailModal } from '../modal';
-import { removeScrollEvent } from '../scrollEvent';
+import { removeScrollEvent } from '../event/scrollEvent';
 
 interface IMovieItemProps {
   id: number;
