+++ src/view/utils/domUtil.ts
@@ -1,3 +1,5 @@
+/* eslint-disable import/prefer-default-export */
+
 export function elementsReplaceWith(targetElements: HTMLElement[], replaceElements: HTMLElement[]) {
   const loopCount = targetElements.length;
   for (let i = 0; i < loopCount; i += 1) {
